+++ packages/shared/CHANGELOG.md
@@ -1,3 +1,29 @@
+## 0.3.1 (2024-10-16)
+
+This was a version bump only for shared to align it with other projects, there were no code changes.
+
+## 0.3.0 (2024-10-16)
+
+
+### 🩹 Fixes
+
+- **shared:** updated searchbar types
+
+## 0.2.0 (2024-10-16)
+
+
+### 🚀 Features
+
+- **shared:** init commit
+
+- **shared:** added data table, searchbar, pagination and error component
+
+- **shared:** added pageHeader, loading, higlightMatches and form footer components
+
+- **shared:** made FormFooter component responsive
+
+- **shared:** added local mode in searchbar
+
 ## 0.1.0 (2024-09-17)